Tests: accept interactive Codex model selection
